// UI/src/app/Apis/ControllerPaths/ControllerPaths.ts
export enum ControllerPaths {
  ROLEMENUITEMS = '/roleMenuItem',
  MENUITEMS = '/menu/item',
  MENU = '/menu',
  ROLE = '/role',
  PROJECT = '/project',
  STORE = '/store',
  STOREINCHARGE = '/storeIncharge',
  USER = '/user',
  ITEMCATEGORY = '/itemcategory',
  BLOCK = '/block',
  COMMON_REFERENCE_TYPE = '/commonreferencedetails/types',
  INDENT = '/indent',
  INDENT_ITEM = '/indent/item',
  ITEMSUBCATEGORY = '/itemsubcategory',
  ITEMSPECIFICATION = '/itemSpecification',
  ITEMUNIT = '/itemUnit',
  WORKFLOWTYPE = '/workflowtype',
  WORKFLOWSTAGE = '/workflowstage',
  LEVEL = '/level',
  UNIT = '/unit',
  HOLIDAY = '/holiday',
  WORKFLOWSERVICE = '/workflowservice',
  UNITTYPE = '/unit/type',
  COMMONREFERENCETYPE = '/commonreferencetype',
  ROLEREFERENCEPERMISSIONTYPE = '/role-permissions',
  COMMONREFERENCEDETAILS = '/commonreferencedetails',
  LEAVEREQUEST = '/leaverequest',
  WEEKOFF = '/weekoff',
  NOTIFICATION = '/notifications',
  STAGE = '/stage',
  PAYMENTPLAN = '/paymentplan',
  CLIENTINVOICEREPORTDATA = 'clientinvoicereportdata',
  REPORT_FIELDS = '/report/fields',
  EXCEL_MAPPING = '/excelmapping',
  Employee = '/employee',
  EmployeeBankingDetails = '/employeeBankDetails',
  ASSETS = '/assets',
  ASSETALLOCATION = '/assetallocation',
  QR = 'QR',
  QRTRANSACTIONDATA = 'qrtransaction',
  QRGENERATOR = 'facility',
  SECURITYPATROL = 'securitypatrol',
  SCHEDULETIME = 'scheduletime',
  ORGANIZATION = 'organization',
  REPORTS = '/reports',
  INVOICE = 'invoicereport',
  INVOICEREPORT = 'taxinvoice',
  QUOTATION = '/quotation',
  QUOTATION_CHARGE = '/charges',
  QUOTATIONITEM = '/quotation/item',
  CUSTOMER = '/customer',
  ISSUES = '/issues',
  ADDRESS = '/address',
  QRREPORT = 'qrreport',
  CUSTOMERCONSUMPTION = 'consumption',
  FACILITYSERVICES = 'facilityservices',
  AGNACUSTOMERCONSUMPTION = 'facilityInvoiceReport',
  BOOKING = '/booking',
  CLIENT = '/client',
  SALETEAM = '/sales/team',
  CLIENTINVOICEREPORT = 'monthlyClientInvoiceReport',
  LOGENTRY = '/logentry',
  DUPLICATELEADHISTORY = 'DUPLICATELEADHISTORY',
  PRIMEACTIVITYCODE = 'primeactivitycode',
  SERVICEGROUP = 'servicegroup',
  SERVICECODE = 'servicecode',
  ACCOUNTENTRY = '/accountentry',
  ACCOUNTAUDITLOG = '/accountauditlog',
  ATTENDANCE = '/attendance',
  SHIFTATTENDANCE = '/shiftBasedAttendance',
  DOCUMENTS = '/documents',
  Notifications = '/notifications',
  ADDITIONALSHIFT = '/additionalshift',
  FINGERPRINTUSERMAPPING = '/fingerprintusermapping',
  EMPLOYEEBANKDETAILS = '/employeeBankDetails',
  SALESAGGREMENTTEMPLATE = '/salesaggrementtemplate',
  SALESAGGREMENTTEMPLATEFIELDS = '/salesAgreementTemplateFields',
  GENERATESALESAGGREMENT = '/generateSalesAgreement',
  MULTILOGIN = '/multilogin',
  CACHE = '/cache',
  CUSTOMERLEGALDOCUMENT = '/customerlegaldocument',
  PROJECTDETAILS = '/projectdetails',

  INVENTORY_RECEIVABLE = '/inventoryreceivables',
  INVENTORY_RECEIVABLE_HISTORY = '/inventoryreceivable/history',
  INVENTORY_RECEIVABLE_ITEM = '/inventoryreceivable/items',
  INVENTORY_RECEIVABLE_STAGE = '/api/inventory-receivable-stage',
  STORE_INVENTORY = '/store/inventory',
  INVENTORY_TRANSFER = '/inventorytransfer',

  QUOTATION_ITEM = '/quotation/item',
  COMMON_REFERENCE_KEY = '/commonreferencedetails/mob',
  FEILDOFFICER_PATROL = '/fieldOfficerPatrol',
  DELETE_FEILDOFFICER_PATROL = '/',
}
// ...
export enum ProcurementControllerPaths {
  PURCHASE_ORDER = '/purchase/order',
  QUOTATION = '/quotation',
  DOCUMENTS = '/documents',
  QUO_TERMS_AND_COND = '/qo/terms/condition',
  STORE_INVENTORY = '/store/inventory',
}